fix: fixed UI placement issue
